Various Changes RE: libvulkan loader
This is a whole bunch of changes related to the loader object and
specifically the libvulkan implementation. The main purpose of this is
refocus the project on my primary goals and not abstract theorizing
about what other clients might need/want/think they want.

The highlights of this are as follows:

- The "plugins" are now "adaptors" which I think makes more sense
- No public command resolution. You aren't going to need it
- There's a good concise test for the loader (which has really limited
  functionality to begin with) now. TDD, etc.