Fixes for Vulkan Definitions
This is two short fixes for vulkandefs.hpp. First, I've namespaced
references to megatech::vulkan::error. This is important in contexts
where the namespace needs to be explicitly resolved (mostly in tests).
Second, I've added non-throwing variants of the
MEGATECH_VULKAN_INTERNAL_BASE_DECLARE_* macros. This is necessary for
loading a Vulkan function in destructors. Unfortunately, because
destructors are non-throwing, there are a variety of error scenarios
where their failure ruins the program's state. Not much to do about it.
Perhaps, I could check for the validity of destructor functions during
construction (i.e., check for vkDestroyInstance right after calling
vkCreateInstance).