Add Consistent Parent and Handle Behavior
This adds consistent parent() and handle() methods on the appropriate
internal objects. Basically, a client should be able to walk up the
ownership chain of an implementation to get at the parent behavior.
Similarly, a client should be able to retrieve the underlying Vulkan
handles to create their own implementation behavior.