Improve Ownership Chain
This improves the ownership relationship between loaders and their
parent vkGetInstanceProcAddr function. The key thing here is that it
unifies the public API. Now, a loader can be discarded safely once an
instance is created. To facilitate this, I've created a new
loader_function_owner type in the internal API. Adaptor implementations
will need to extend this type and provide at least a constructor and
destructor (if appropriate).