Improve Physical Device Validation
This moves physical device validation out of the
physical_device_validator type and into
physical_device_description_impl. There's a moderate increase in
complexity for the physical device description, but doing this will help
me eliminate a couple types (physical_device_validator, and
device_description). It also limits the complexity of extending the
device requirements in an adaptor. Devices will now be much more
configurable with much less work on the part of an implementor. Of
course, there will need to be a good amount of documentation to go with
this.

- Moves is_valid() to physical_device_description_impl
- Turns physical_device_validator into a stub. This will be removed in a
  later commit.
- Updates physical_device_list to use the is_valid() method.