Add Assertions
This adds Megatech-Assertions to the build. I've submoduled it, but it's
also a submodule of Megatech-Vulkan Dispatch. Meson should select the
higher version (iirc). I've tried to favor thoroughness over performance
with the assertions. A lot of things get checked two or three times as
they pass through layers of the API. In my opinion, this is ok. This
ensures each layer isn't fudging anything. It's probably overkill
though.

I've made a few code changes alongside this. The most major one is a few
more operators for bitmasks.