Simplify Adaptor Implementation
The main thing here is simplifying the implementation of adaptors.
Unlike previously, an adaptor is only needed to create the loader.
Adaptors that need to extend the instance can do so by extending
instance_impl and overriding the resolver method in loader_impl.
Similarly, adaptors that need to extend the
physical_device_description_impl will need to override the
resolver method in instance_impl. Basically, only implement as much as
you need.

Several objects are removed (all of the allocator stuff), but if I
really hate the resolver stuff I might swap back to it.

This also updates documentation and makes changes to the build scripts.
There's probably some code cleanup too.